chore: apply clippy fixes and code formatting (T090-T092)
Phase 9 cleanup tasks:
- T090: Removed debug console.log in ScoreViewer.tsx
- T091: Fixed clippy.toml config and applied auto-fixes
- T092: Ran cargo fmt --all for consistent formatting

Changes:
- Fixed .clippy.toml (removed invalid [lints] sections)
- Applied clippy --fix (unused variables, unnecessary casts, range contains)
- Formatted all Rust code with cargo fmt
- Removed temporary debug logging from frontend

Remaining clippy warnings are style suggestions (too many args, etc.)
242 tests still passing after formatting.

+++ backend/tests/api_integration_test.rs
@@ -728,7 +728,7 @@ async fn test_polyphonic_chord_10_notes() {
             "Failed to add note with pitch {}",
             pitch
         );
-        assert_eq!(body["pitch"], *pitch as i64);
+        assert_eq!(body["pitch"], *pitch);
         assert_eq!(body["start_tick"], start_tick);
     }
